Address review feedback on testing infrastructure
- Remove routine_name from RoutineConvKey for singleton semantics
- Add get_or_create_in_cache helper to eliminate duplication
- Extract swap_listener to share logic between restart methods
- Fix restart_with_listener to call local_addr before mutating state
- Handle mutex poisoning gracefully in next_synthetic_uuid
- Replace unimplemented! in StubLlm with deterministic Result returns
- Add docs/testing-abstractions.md with testing guide
- Add comment explaining _stream lifecycle in sighup_reload.rs
- Move webhook_server tests to tests/webhook_server.rs
