+++ tests/webhook_server.rs
@@ -1,0 +1,164 @@
+//! Integration tests for WebhookServer.
+
+use std::net::SocketAddr;
+use std::net::TcpListener as StdTcpListener;
+
+use axum::Json;
+use axum::Router;
+use rstest::{fixture, rstest};
+use serde_json::json;
+
+use ironclaw::channels::{WebhookServer, WebhookServerConfig};
+
+/// A started webhook server with a `/health` route and a pre-built client.
+struct StartedWebhookServer {
+    server: WebhookServer,
+    addr: SocketAddr,
+    client: reqwest::Client,
+}
+
+/// Binds an ephemeral port, creates a [`WebhookServer`] with a `/health`
+/// route, starts the server on the already-bound listener, and returns the
+/// address and a client.
+#[fixture]
+async fn started_webhook_server()
+-> Result<StartedWebhookServer, Box<dyn std::error::Error + Send + Sync>> {
+    let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await?;
+    let addr = listener.local_addr()?;
+    let mut server = WebhookServer::new(WebhookServerConfig { addr });
+    server.add_routes(Router::new().route(
+        "/health",
+        axum::routing::get(|| async { Json(json!({"status": "ok"})) }),
+    ));
+    server.start_with_listener(listener).await?;
+    Ok(StartedWebhookServer {
+        server,
+        addr,
+        client: reqwest::Client::new(),
+    })
+}
+
+#[rstest]
+#[tokio::test]
+async fn test_restart_with_addr_rebinds_listener(
+    #[future] started_webhook_server: Result<
+        StartedWebhookServer,
+        Box<dyn std::error::Error + Send + Sync>,
+    >,
+) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
+    let StartedWebhookServer {
+        mut server,
+        addr: addr1,
+        client,
+    } = started_webhook_server.await?;
+
+    assert_eq!(
+        server.current_addr(),
+        addr1,
+        "Server should be bound to initial address"
+    );
+
+    let response = client
+        .get(format!("http://{}/health", addr1))
+        .send()
+        .await?;
+    assert_eq!(
+        response.status(),
+        200,
+        "First server should respond to health check"
+    );
+
+    // Find a second available port and restart
+    let port2 = {
+        let listener = StdTcpListener::bind("127.0.0.1:0")?;
+        listener.local_addr()?.port()
+    };
+    let addr2: SocketAddr = format!("127.0.0.1:{}", port2).parse()?;
+
+    server.restart_with_addr(addr2).await?;
+
+    assert_eq!(
+        server.current_addr(),
+        addr2,
+        "Server address should be updated after restart"
+    );
+    assert_ne!(
+        addr1, addr2,
+        "Address should change after restart_with_addr"
+    );
+
+    let response = client
+        .get(format!("http://{}/health", addr2))
+        .send()
+        .await?;
+    assert_eq!(
+        response.status(),
+        200,
+        "Restarted server should respond to health check on new address"
+    );
+
+    let old_result = tokio::time::timeout(
+        std::time::Duration::from_millis(200),
+        client.get(format!("http://{}/health", addr1)).send(),
+    )
+    .await;
+    assert!(
+        old_result.is_err() || old_result.ok().and_then(|r| r.ok()).is_none(),
+        "Old address should not respond after server restarts"
+    );
+
+    server.shutdown().await;
+    Ok(())
+}
+
+#[rstest]
+#[tokio::test]
+async fn test_restart_with_addr_rollback_on_bind_failure(
+    #[future] started_webhook_server: Result<
+        StartedWebhookServer,
+        Box<dyn std::error::Error + Send + Sync>,
+    >,
+) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
+    let StartedWebhookServer {
+        mut server,
+        addr: addr1,
+        client,
+    } = started_webhook_server.await?;
+
+    let response = client
+        .get(format!("http://{}/health", addr1))
+        .send()
+        .await?;
+    assert_eq!(response.status(), 200, "Server should be listening");
+
+    // Occupy a second port so the restart bind fails deterministically.
+    let occupied_listener = StdTcpListener::bind("127.0.0.1:0")?;
+    let conflict_addr = occupied_listener.local_addr()?;
+
+    let result = server.restart_with_addr(conflict_addr).await;
+    assert!(
+        result.is_err(),
+        "Restart with already-bound address should fail"
+    );
+
+    drop(occupied_listener);
+
+    let response = client
+        .get(format!("http://{}/health", addr1))
+        .send()
+        .await?;
+    assert_eq!(
+        response.status(),
+        200,
+        "Old listener should still be running after failed restart"
+    );
+
+    assert_eq!(
+        server.current_addr(),
+        addr1,
+        "Server address should be restored after failed restart"
+    );
+
+    server.shutdown().await;
+    Ok(())
+}
